+++ CHANGES.md
@@ -30,6 +30,7 @@
 - The "Duplicate" event context menu item now has "..." at the end.
 - Changed the default background color for the tooltips shown through the Calendar (slightly lighter now).
 - The DatePicker mode now supports smaller mobile screens (and increased the default width to 330px).
+- Added new setting "addYearButtonsInDatePickerMode" (defaults to false) which adds year-jumping buttons to DatePicker title bar.
 - 
 
 ### Version 1.8.7: